Ready to deploy

+++ controllers/categoryController.js
@@ -1,9 +1,9 @@
-const { Category } = require('../models/index');
+const { Category, Product } = require('../models/index');
 
 class CategoryController {
   static async getCategories(req, res, next) {
     try {
-      const categories = await Category.findAll();
+      const categories = await Category.findAll({order: [['name', 'ASC']]});
       res.status(200).json(categories);
     } catch(err) {
       next(err);
@@ -22,6 +22,14 @@ class CategoryController {
 
   static async deleteCategory(req, res, next) {
     try {
+      const product = await Product.findOne({
+        where: {
+          CategoryId: req.params.id
+        }
+      })
+      if(product) {
+        throw {name: 'CategoryHaveProductsInit'}
+      }
       const destroyed = await Category.destroy({
         where: {
           id: req.params.id
